third_party/WebKit/LayoutTests/webaudio/internals/cycle-connection-gc.html

<!DOCTYPE html>
<html>
  <head>
    <title>
      cycle-connection-gc.html
    </title>
    <script src="../../resources/js-test.js"></script>
  </head>
  <body>
    <script id="layout-test-code">
      description('Cycles of AudioNode connections should be collected.');
      let context = new OfflineAudioContext(2, 44100, 44100);
      gc();
      let initialCount = internals.audioHandlerCount();
      createCycle();
      debug('A cycle was created:');
      shouldBeTrue('internals.audioHandlerCount() > initialCount');
      gc();
      debug('GC happened:');
      shouldBe('internals.audioHandlerCount()', 'initialCount');

      function createCycle() {
        let source = context.createBufferSource();
        // Can't use DelayNode here because they now have a tail time that keeps
        // the node handler alive to process the tail.  Use GainNodes instead
        // because they have no tail time.  (Strictly speaking, there should be
        // a delay node in the cycle, but we don't enforce this.)
        let gain1 = context.createGain();
        let gain2 = context.createGain();
        source.connect(gain1);
        gain1.connect(gain2);
        gain2.connect(gain1);
        gain1.connect(context.destination);
      }
    </script>
  </body>
</html>
